fix: rename my/components and frame dir
