+++ packages/web/src/app/my/activity-certificate/[id]/page.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 
-import MyCertificateDetailFrame from "@sparcs-clubs/web/features/my/activity-certificate/frame/MyCertificateDetailFrame";
+import MyCertificateDetailFrame from "@sparcs-clubs/web/features/my/activity-certificate/frames/MyCertificateDetailFrame";
 
 const MyAcfDetail = () => <MyCertificateDetailFrame />;
