+++ packages/web/src/app/my/page.tsx
@@ -8,11 +8,11 @@ import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import PageHead from "@sparcs-clubs/web/common/components/PageHead";
 import useGetUserProfile from "@sparcs-clubs/web/common/services/getUserProfile";
-import MyChangeDivisionPresident from "@sparcs-clubs/web/features/my/component/MyChangeDivisionPresident";
-import MyChangeRepresentative from "@sparcs-clubs/web/features/my/component/MyChangeRepresentative";
-import MyClubFrame from "@sparcs-clubs/web/features/my/frame/MyClubFrame";
-import MyInfoFrame from "@sparcs-clubs/web/features/my/frame/MyInfoFrame";
-import MyServiceFrame from "@sparcs-clubs/web/features/my/frame/MyServiceFrame";
+import MyChangeDivisionPresident from "@sparcs-clubs/web/features/my/components/MyChangeDivisionPresident";
+import MyChangeRepresentative from "@sparcs-clubs/web/features/my/components/MyChangeRepresentative";
+import MyClubFrame from "@sparcs-clubs/web/features/my/frames/MyClubFrame";
+import MyInfoFrame from "@sparcs-clubs/web/features/my/frames/MyInfoFrame";
+import MyServiceFrame from "@sparcs-clubs/web/features/my/frames/MyServiceFrame";
 import { useGetMyDelegateRequest } from "@sparcs-clubs/web/features/my/services/getMyDelegateRequest";
 
 const My: React.FC = () => {
