+++ packages/web/src/app/my/printing-business/[id]/page.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 
-import MyPrintingDetailFrame from "@sparcs-clubs/web/features/my/printing-business/frame/MyPrintingDetailFrame";
+import MyPrintingDetailFrame from "@sparcs-clubs/web/features/my/printing-business/frames/MyPrintingDetailFrame";
 
 const MyPrintingDetail = () => <MyPrintingDetailFrame />;
